Move gotron to root dir
