+++ browser_window_instance_methods.go
