+++ browser_window_test.go
