+++ gbw-doc/BrowserWindow.md
