+++ gotron_methods_pack.go
@@ -1,4 +1,4 @@
-// +build gotronbrowserwindowprod
+// +build gotronpack
 
 package gotron
