+++ templates/app/main.js
